Add files via upload
How It Works
Data Preprocessing: The data is loaded, and the content is combined from the title and author fields. Text preprocessing is performed to clean the text and prepare it for analysis.
Feature Extraction: The text data is converted into numerical format using TF-IDF vectorization.
Model Training: The dataset is split into training and testing sets. A Random Forest Classifier model is trained on the training data.
Prediction: The model predicts whether new news articles are real or fake based on the learned patterns.